+++ src/main.rs
@@ -52,8 +52,8 @@ lazy_static::lazy_static! {
     /// Path to `debug.log` file. Standard is `$HOME/.local/share/honkers-railway-launcher/debug.log`
     pub static ref DEBUG_FILE: PathBuf = LAUNCHER_FOLDER.join("debug.log");
 
-    /// Path to `background` file. Standard is `$HOME/.cache/honkers-railway-launcher/background`
-    pub static ref BACKGROUND_FILE: PathBuf = CACHE_FOLDER.join("background");
+    /// Path to `background` file. Standard is `$HOME/.local/share/honkers-railway-launcher/background`
+    pub static ref BACKGROUND_FILE: PathBuf = LAUNCHER_FOLDER.join("background");
 
     /// Path to `.keep-background` file. Used to mark launcher that it shouldn't update background picture
     ///