+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.2] - 18.06.2023
+
 ### Added
 
 - Added deletion of old patch files (just in case)
@@ -110,8 +112,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <br>
 
-[unreleased]: https://github.com/an-anime-team/the-honkers-railway-launcher/compare/1.2.1...next
-[1.2.0]: https://github.com/an-anime-team/the-honkers-railway-launcher/compare/1.2.0...1.2.1
+[unreleased]: https://github.com/an-anime-team/the-honkers-railway-launcher/compare/1.2.2...next
+[1.2.2]: https://github.com/an-anime-team/the-honkers-railway-launcher/compare/1.2.1...1.2.2
+[1.2.1]: https://github.com/an-anime-team/the-honkers-railway-launcher/compare/1.2.0...1.2.1
 [1.2.0]: https://github.com/an-anime-team/the-honkers-railway-launcher/compare/1.1.0...1.2.0
 [1.1.0]: https://github.com/an-anime-team/the-honkers-railway-launcher/compare/1.0.0...1.1.0
 [1.0.0]: https://github.com/an-anime-team/the-honkers-railway-launcher/releases/tag/1.0.0